# Bump deps and Elm 0.19 JS optimization