--- a/build-utils/webpack.production.js
+++ b/build-utils/webpack.production.js
@@ -17,7 +17,12 @@ module.exports = () => ({
       {
         test: /\.elm$/,
         exclude: [/elm-stuff/, /node_modules/],
-        use: 'elm-webpack-loader'
+        use: {
+          loader: 'elm-webpack-loader',
+          options: {
+            optimize: true,
+          },
+        },
       },
       {
         test: /\.(sa|sc|c)ss$/,
@@ -29,6 +34,31 @@ module.exports = () => ({
         ],
       }
     ]
+  },
+
+  optimization: {
+    minimizer: [
+      // https://elm-lang.org/0.19.0/optimize
+      new UglifyJsPlugin({
+        cache: true,
+        uglifyOptions: {
+          compress: {
+            pure_funcs: ['F2','F3','F4','F5','F6','F7','F8','F9','A2','A3','A4','A5','A6','A7','A8','A9'],
+            pure_getters: true,
+            keep_fargs: false,
+            unsafe_comps: true,
+            unsafe: true,
+          },
+          mangle: false,
+        },
+      }),
+      new UglifyJsPlugin({
+        cache: true,
+        uglifyOptions: {
+          mangle: true,
+        },
+      }),
+    ],
   },
 
   plugins: [
@@ -48,21 +78,6 @@ module.exports = () => ({
     ]),
 
     new ImageminPlugin({ test: /\.(jpe?g|png|gif|svg)$/i }),
-
-    new UglifyJsPlugin({
-      cache: true,
-      parallel: true,
-      uglifyOptions: {
-        compress: {
-          pure_funcs: ['F2','F3','F4','F5','F6','F7','F8','F9','A2','A3','A4','A5','A6','A7','A8','A9'],
-          pure_getters: true,
-          keep_fargs: false,
-          unsafe_comps: true,
-          unsafe: true,
-          passes: 3,
-        },
-      },
-    }),
 
     new OptimizeCSSAssetsPlugin()
   ]