Bump deps and better UglifyJS options for optmization.

Following https://elm-lang.org/0.19.0/optimize

Currently there is no way to choose what get's minimized before the output file (https://github.com/webpack-contrib/uglifyjs-webpack-plugin/issues/21). There are workarounds which I choose not to use. Closes #41 

There is also a way to optimize Elm in 1 pass instead of 2 (no webpack support). More information: https://github.com/rtfeldman/elm-spa-example/pull/54. The webpack loader doesn't accept the `rename` option

## Before
#### Elm + Bootstrap
```sh
Version: webpack 4.24.0
Time: 2725ms
                                      Asset       Size  Chunks             Chunk Names
               main.4e9361d0100827c946a3.js    190 KiB       0  [emitted]  main
            main.4e9361d0100827c946a3.js.gz   61.3 KiB          [emitted]
```

#### Elm
```sh
Version: webpack 4.24.0
Time: 7253ms
                                      Asset       Size  Chunks             Chunk Names
               main.e66a90029fcb2bfeac6a.js   35.8 KiB       0  [emitted]  main
            main.e66a90029fcb2bfeac6a.js.gz   13.2 KiB          [emitted]
```

## After
#### Elm + Bootstrap
```sh
Version: webpack 4.24.0
Time: 8560ms
                                      Asset       Size  Chunks             Chunk Names
               main.97952e0cdad5b5f56332.js    181 KiB       0  [emitted]  main
            main.97952e0cdad5b5f56332.js.gz     59 KiB          [emitted]
```

#### Elm
```sh
Version: webpack 4.24.0
Time: 4607ms
                                      Asset       Size  Chunks             Chunk Names
               main.649703dca62fa38ddf07.js   27.4 KiB       0  [emitted]  main
            main.649703dca62fa38ddf07.js.gz   10.4 KiB          [emitted]
```